Move DA to it's own module
