+++ crates/sequencer/src/lib.rs
@@ -22,6 +22,7 @@ use tokio::sync::mpsc::unbounded_channel;
 
 mod commitment;
 mod compat;
+mod da;
 pub mod db_migrations;
 mod db_provider;
 mod deposit_data_mempool;
